requirements.txt
flask==2.3.0
tensorflow==2.13.0
numpy==1.24.3
opencv-python==4.8.0.74
pillow==10.0.0
werkzeug==2.3.0